refactor(ui): extract planned row text projection
