+++ src/opentui/HunkDiffBody.tsx
@@ -2,7 +2,8 @@ import { useMemo } from "react";
 import { DEFAULT_TAB_WIDTH } from "../core/run/tabWidth";
 import { findMaxLineNumber } from "../ui/diff/codeColumns";
 import { buildSplitRows, buildStackRows } from "../ui/diff/diffRows";
-import { diffMessage, DiffRowView, fitText } from "../ui/diff/renderRows";
+import { diffMessage, fitText } from "../ui/diff/plannedRowText";
+import { DiffRowView } from "../ui/diff/renderRows";
 import { useHighlightedDiff } from "../ui/diff/useHighlightedDiff";
 import { resolveTheme } from "../ui/themes";
 import { toInternalDiffFile } from "./model";
