+++ src/ui/diff/renderRows.tsx
@@ -5,7 +5,6 @@ import {
   type MouseEvent as TuiMouseEvent,
   type TextChunk,
 } from "@opentui/core";
-import type { DiffFile } from "../../core/changeset/model";
 import type { UserNoteLineTarget } from "../../core/liveComments";
 import type { AppTheme } from "../themes";
 import { CODE_ROW_ADD_NOTE_BADGE_TEXT, CODE_ROW_ADD_NOTE_BADGE_WIDTH } from "./codeRowAffordance";
@@ -16,7 +15,6 @@ import {
   type CodeRowLayoutPlan,
   type PlannedDiffReviewRow,
 } from "./codeRowLayout";
-import { reviewEmptyDiffReason, type ReviewEmptyDiffReason } from "../../core/review/document";
 import { reviewGapId } from "../../core/review/expansion";
 import type { DiffRow, RenderSpan, SplitLineCell, StackLineCell } from "./diffRows";
 import {
@@ -39,17 +37,9 @@ import {
   stackGutterText,
   stackRailColor,
 } from "./rowStyle";
-import { type PlannedReviewRow } from "./reviewRenderPlan";
-import { inlineNoteTitle } from "../components/panes/AgentInlineNote";
-import { wrapText } from "../lib/text";
-import { sanitizeTerminalLine, sanitizeTerminalSpans } from "../../lib/terminalText";
-import {
-  measureSanitizedTextWidth,
-  measureTextWidth,
-  padText as padTextByWidth,
-  sliceSanitizedTextByWidth,
-  sliceTextByWidth,
-} from "../lib/text";
+import { sanitizeTerminalSpans } from "../../lib/terminalText";
+import { measureTextWidth, sliceTextByWidth } from "../lib/text";
+import { fitText } from "./plannedRowText";
 import { sliceSpansWindow, wrapSpans } from "./styledSpanLayout";
 import type { CopySelectedRowRange } from "../lib/diffSpatial";
 import type { CursorLine } from "../../core/run/commandInputs";
@@ -87,24 +77,6 @@ interface CellPrefix {
 
 /** Column span covering a row's whole content column, in the global columns selection uses. */
 const FULL_ROW_COL_RANGE: CopySelectedRowRange = { startCol: 0, endCol: Number.MAX_SAFE_INTEGER };
-
-/** Clamp a label to one terminal row with an ellipsis. */
-export function fitText(text: string, width: number) {
-  const safeText = sanitizeTerminalLine(text);
-  if (width <= 0) {
-    return "";
-  }
-
-  if (measureSanitizedTextWidth(safeText) <= width) {
-    return safeText;
-  }
-
-  if (width === 1) {
-    return "…";
-  }
-
-  return `${sliceSanitizedTextByWidth(safeText, 0, width - 1).text}…`;
-}
 
 const marker = diffRailMarker;
 const styledTextColorCache = new Map<string, ReturnType<typeof parseColor>>();
@@ -623,382 +595,6 @@ function buildWrappedStackCell(
   } satisfies WrappedCellLayout;
 }
 
-/** Convert a list of spans to fixed-width plain text while preserving logical clipping. */
-function spansToPlainText(spans: RenderSpan[], width: number, horizontalOffset = 0) {
-  if (width <= 0) {
-    return "";
-  }
-
-  const visible = sliceTextByWidth(
-    sanitizeTerminalSpans(spans)
-      .map((span) => span.text)
-      .join(""),
-    Math.max(0, horizontalOffset),
-    width,
-  );
-
-  return padTextByWidth(visible.text, Math.max(0, width));
-}
-
-/** Flatten styled spans to their visible text content. */
-function spansText(spans: RenderSpan[]) {
-  return sanitizeTerminalSpans(spans)
-    .map((span) => span.text)
-    .join("");
-}
-
-/** Return one cell's code text without rail, gutter, sign, or line-number decorations. */
-function cellCodeText(spans: RenderSpan[], horizontalOffset = 0) {
-  return sliceTextByWidth(spansText(spans), Math.max(0, horizontalOffset), Number.MAX_SAFE_INTEGER)
-    .text;
-}
-
-function buildPlainSplitCellLines(
-  cell: SplitLineCell,
-  geometry: CodeCellLayoutPlan,
-  lineNumberDigits: number,
-  showLineNumbers: boolean,
-  theme: AppTheme,
-  wrapLines: boolean,
-  codeHorizontalOffset = 0,
-) {
-  if (!wrapLines) {
-    const gutterText = splitGutterText(cell, lineNumberDigits, showLineNumbers).padEnd(
-      geometry.gutterWidth,
-    );
-
-    return [
-      {
-        contentWidth: geometry.contentWidth,
-        gutterWidth: geometry.gutterWidth,
-        spansText:
-          gutterText + spansToPlainText(cell.spans, geometry.contentWidth, codeHorizontalOffset),
-      },
-    ];
-  }
-
-  const layout = buildWrappedSplitCell(cell, geometry, lineNumberDigits, showLineNumbers, theme);
-
-  // Mirror the TUI renderer, which does not apply horizontal scrolling to wrapped rows.
-  // Keeping the plain-text path aligned avoids visual/clipboard drift.
-  return layout.lines.map((line) => ({
-    contentWidth: layout.contentWidth,
-    gutterWidth: layout.gutterWidth,
-    spansText: line.gutterText + spansToPlainText(line.spans, layout.contentWidth),
-  }));
-}
-
-function buildPlainStackCellLines(
-  cell: StackLineCell,
-  geometry: CodeCellLayoutPlan,
-  lineNumberDigits: number,
-  showLineNumbers: boolean,
-  theme: AppTheme,
-  wrapLines: boolean,
-  codeHorizontalOffset = 0,
-) {
-  if (!wrapLines) {
-    const gutterText = stackGutterText(cell, lineNumberDigits, showLineNumbers).padEnd(
-      geometry.gutterWidth,
-    );
-
-    return [
-      {
-        contentWidth: geometry.contentWidth,
-        gutterWidth: geometry.gutterWidth,
-        spansText:
-          gutterText + spansToPlainText(cell.spans, geometry.contentWidth, codeHorizontalOffset),
-      },
-    ];
-  }
-
-  const layout = buildWrappedStackCell(cell, geometry, lineNumberDigits, showLineNumbers, theme);
-
-  // Mirror the TUI renderer, which does not apply horizontal scrolling to wrapped rows.
-  return layout.lines.map((line) => ({
-    contentWidth: layout.contentWidth,
-    gutterWidth: layout.gutterWidth,
-    spansText: line.gutterText + spansToPlainText(line.spans, layout.contentWidth),
-  }));
-}
-
-/** Render the marker + label that hunk-header and collapsed rows share in plain-text form. */
-function renderHeaderRowText(text: string, width: number) {
-  const label = fitText(text, Math.max(0, width - 1));
-  return marker() + padTextByWidth(label, Math.max(0, width - 1));
-}
-
-interface PlannedRowTextOptions {
-  width: number;
-  lineNumberDigits: number;
-  showLineNumbers: boolean;
-  showHunkHeaders: boolean;
-  wrapLines: boolean;
-  codeHorizontalOffset: number;
-  theme: AppTheme;
-  reserveAddNoteColumn?: boolean;
-  showAddNoteBadge?: boolean;
-  // When set, split-line rows produce text only for this side. Stack rows ignore the filter.
-  side?: "left" | "right";
-}
-
-/** Render one or more decorated plain-text lines for one planned row. */
-export function renderDecoratedPlannedRowText(
-  row: PlannedReviewRow,
-  options: PlannedRowTextOptions,
-) {
-  const {
-    width,
-    lineNumberDigits,
-    showLineNumbers,
-    showHunkHeaders,
-    wrapLines,
-    codeHorizontalOffset,
-    theme,
-    side,
-  } = options;
-
-  if (width <= 0) {
-    return [];
-  }
-
-  if (row.kind === "inline-note") {
-    const title = inlineNoteTitle(row.annotation, row.noteIndex, row.noteCount);
-    const summaryLines = wrapText(row.annotation.summary ?? "", width).map((line) =>
-      fitText(line, width),
-    );
-    const rationaleLines = row.annotation.rationale
-      ? wrapText(row.annotation.rationale, width).map((line) => fitText(line, width))
-      : [];
-    return [fitText(title, width), ...summaryLines, ...rationaleLines];
-  }
-
-  const preparedRow = row.row;
-
-  if (preparedRow.type === "hunk-header") {
-    return showHunkHeaders ? [renderHeaderRowText(preparedRow.text, width)] : [];
-  }
-
-  if (preparedRow.type === "collapsed") {
-    return [renderHeaderRowText(`··· ${preparedRow.text} ···`, width)];
-  }
-
-  if (preparedRow.type === "split-line") {
-    const codeLayout = planCodeRowLayout(row, options) as Extract<
-      CodeRowLayoutPlan,
-      { kind: "split" }
-    >;
-    const guideOnOldSide = codeLayout.noteGuideSide === "old";
-    const guideOnNewSide = codeLayout.noteGuideSide === "new";
-    const leftPrefix = guideOnOldSide ? "│" : marker();
-    const rightPrefix = "▌";
-
-    const leftCell = buildPlainSplitCellLines(
-      preparedRow.left,
-      codeLayout.left,
-      lineNumberDigits,
-      showLineNumbers,
-      theme,
-      wrapLines,
-      codeHorizontalOffset,
-    );
-    const rightCell = buildPlainSplitCellLines(
-      preparedRow.right,
-      codeLayout.right,
-      lineNumberDigits,
-      showLineNumbers,
-      theme,
-      wrapLines,
-      codeHorizontalOffset,
-    );
-
-    const visualLineCount = Math.max(leftCell.length, rightCell.length);
-    return Array.from({ length: visualLineCount }, (_, index) => {
-      const leftLine = leftCell[index] ?? {
-        gutterWidth: codeLayout.left.gutterWidth,
-        contentWidth: codeLayout.left.contentWidth,
-        spansText: " ".repeat(Math.max(0, codeLayout.left.width - codeLayout.left.prefixWidth)),
-      };
-      const rightLine = rightCell[index] ?? {
-        gutterWidth: codeLayout.right.gutterWidth,
-        contentWidth: codeLayout.right.contentWidth,
-        spansText: " ".repeat(Math.max(0, codeLayout.right.width - codeLayout.right.prefixWidth)),
-      };
-      const normalizedLeft = padTextByWidth(
-        `${leftPrefix}${leftLine.spansText}`,
-        codeLayout.left.width,
-      );
-      const normalizedRight = padTextByWidth(
-        `${rightPrefix}${rightLine.spansText}`,
-        codeLayout.right.width,
-      );
-
-      if (side === "left") {
-        return normalizedLeft;
-      }
-      if (side === "right") {
-        return `${normalizedRight}${guideOnNewSide ? "│" : ""}`;
-      }
-
-      return `${normalizedLeft}${normalizedRight}${guideOnNewSide ? "│" : ""}`;
-    });
-  }
-
-  if (preparedRow.type !== "stack-line") {
-    return [];
-  }
-
-  const codeLayout = planCodeRowLayout(row, options) as Extract<
-    CodeRowLayoutPlan,
-    { kind: "stack" }
-  >;
-  const guideOnOldSide = codeLayout.noteGuideSide === "old";
-  const guideOnNewSide = codeLayout.noteGuideSide === "new";
-  const prefix = guideOnOldSide ? "│" : marker();
-  const cellLines = buildPlainStackCellLines(
-    preparedRow.cell,
-    codeLayout.cell,
-    lineNumberDigits,
-    showLineNumbers,
-    theme,
-    wrapLines,
-    codeHorizontalOffset,
-  );
-
-  return cellLines.map((line) => {
-    const visibleLine = `${prefix}${line.spansText}`;
-    const normalized = padTextByWidth(visibleLine, Math.max(1, codeLayout.cell.width));
-    return `${normalized}${guideOnNewSide ? "│" : ""}`;
-  });
-}
-
-/**
- * Render only code content for one planned row, excluding gutters, headers, notes, and filenames.
- *
- * Split context rows (left.text === right.text) are deduplicated to a single line because both
- * sides of an unchanged context row carry identical text and shipping it twice in the clipboard
- * would be noise. If the renderer ever distinguishes left/right context via styling that bleeds
- * into the span text itself, this dedupe should be revisited.
- */
-export function renderCodeOnlyPlannedRowText(
-  row: PlannedReviewRow,
-  options: PlannedRowTextOptions,
-) {
-  const { width, lineNumberDigits, showLineNumbers, wrapLines, codeHorizontalOffset, theme, side } =
-    options;
-
-  if (width <= 0 || row.kind !== "diff-row") {
-    return [];
-  }
-
-  const preparedRow = row.row;
-  if (preparedRow.type === "hunk-header" || preparedRow.type === "collapsed") {
-    return [];
-  }
-
-  if (preparedRow.type === "stack-line") {
-    if (!wrapLines) {
-      return [cellCodeText(preparedRow.cell.spans, codeHorizontalOffset)].filter(Boolean);
-    }
-
-    const codeLayout = planCodeRowLayout(row, options) as Extract<
-      CodeRowLayoutPlan,
-      { kind: "stack" }
-    >;
-    return buildWrappedStackCell(
-      preparedRow.cell,
-      codeLayout.cell,
-      lineNumberDigits,
-      showLineNumbers,
-      theme,
-    )
-      .lines.map((line) => spansText(line.spans))
-      .filter(Boolean);
-  }
-
-  if (preparedRow.type !== "split-line") {
-    return [];
-  }
-
-  if (!wrapLines) {
-    const leftText =
-      preparedRow.left.kind === "empty"
-        ? ""
-        : cellCodeText(preparedRow.left.spans, codeHorizontalOffset);
-    const rightText =
-      preparedRow.right.kind === "empty"
-        ? ""
-        : cellCodeText(preparedRow.right.spans, codeHorizontalOffset);
-
-    if (side === "left") {
-      return [leftText].filter(Boolean);
-    }
-    if (side === "right") {
-      return [rightText].filter(Boolean);
-    }
-
-    if (leftText && rightText && leftText === rightText) {
-      return [leftText];
-    }
-
-    return [leftText, rightText].filter(Boolean);
-  }
-
-  const codeLayout = planCodeRowLayout(row, options) as Extract<
-    CodeRowLayoutPlan,
-    { kind: "split" }
-  >;
-  const leftLayout = buildWrappedSplitCell(
-    preparedRow.left,
-    codeLayout.left,
-    lineNumberDigits,
-    showLineNumbers,
-    theme,
-  );
-  const rightLayout = buildWrappedSplitCell(
-    preparedRow.right,
-    codeLayout.right,
-    lineNumberDigits,
-    showLineNumbers,
-    theme,
-  );
-  const visualLineCount = Math.max(leftLayout.lines.length, rightLayout.lines.length);
-  const lines: string[] = [];
-
-  for (let index = 0; index < visualLineCount; index += 1) {
-    const leftText =
-      preparedRow.left.kind === "empty" ? "" : spansText(leftLayout.lines[index]?.spans ?? []);
-    const rightText =
-      preparedRow.right.kind === "empty" ? "" : spansText(rightLayout.lines[index]?.spans ?? []);
-
-    if (side === "left") {
-      if (leftText) {
-        lines.push(leftText);
-      }
-      continue;
-    }
-    if (side === "right") {
-      if (rightText) {
-        lines.push(rightText);
-      }
-      continue;
-    }
-
-    if (leftText && rightText && leftText === rightText) {
-      lines.push(leftText);
-    } else {
-      if (leftText) {
-        lines.push(leftText);
-      }
-      if (rightText) {
-        lines.push(rightText);
-      }
-    }
-  }
-
-  return lines;
-}
-
 /**
  * Apply a highlight blend to a cell palette's gutter bg only.
  *
@@ -1356,27 +952,6 @@ function renderWrappedStackCellLine(
       )}
     </>
   );
-}
-
-/** Review-stream wording for each shared reason a file renders no diff rows. */
-export const DIFF_MESSAGES: Record<ReviewEmptyDiffReason, string> = {
-  "rename-only": "No textual hunks. This change only renames the file.",
-  binary: "Binary file skipped",
-  "too-large": "File too large to render automatically.",
-  "new-file": "No textual hunks. The file is marked as new.",
-  "deleted-file": "No textual hunks. The file is marked as deleted.",
-  "no-hunks": "No textual hunks to render for this file.",
-};
-
-/** Explain why a file still appears in the review stream even when it has no textual hunks. */
-export function diffMessage(file: DiffFile) {
-  return DIFF_MESSAGES[
-    reviewEmptyDiffReason({
-      changeKind: file.metadata.type,
-      binary: Boolean(file.isBinary),
-      tooLarge: Boolean(file.isTooLarge),
-    })
-  ];
 }
 
 /** Build the rendered label text for one collapsed gap row. */
