+++ src/ui/lib/agentAnnotations.ts
@@ -2,6 +2,7 @@ import type { Hunk } from "@pierre/diffs";
 import type { DiffFile } from "../../core/changeset/model";
 import type { ReviewNoteSource } from "../../core/run/commandInputs";
 import type { AgentAnnotation } from "../../extension-api/types";
+import { sanitizeTerminalLine } from "../../lib/terminalText";
 import { reviewAnnotationOverlapsHunk } from "../../core/review/annotations";
 import { resolveReviewNoteAnchor, reviewGapOwnerHunkIndex } from "../../core/review/anchors";
 import type { ReviewHunkSpan } from "../../core/review/geometry";
@@ -34,6 +35,18 @@ export interface VisibleAgentNote {
 export interface AnnotationAnchor {
   side: "old" | "new";
   lineNumber: number;
+}
+
+/** Build the source and author label shown for one inline note. */
+export function inlineNoteTitle(annotation: AgentAnnotation, noteIndex: number, noteCount: number) {
+  if (annotation.source === "user-draft") {
+    return "Draft note";
+  }
+
+  const source = reviewNoteSource(annotation);
+  const author = sanitizeTerminalLine(annotation.author?.trim() ?? "");
+  const label = source === "user" ? "Your note" : author ? `${author} note` : "Agent note";
+  return noteCount > 1 ? `${label} ${noteIndex + 1}/${noteCount}` : label;
 }
 
 /** Resolve the user-facing source for one inline note annotation. */
